+++ examples/manual/dates/dateserial.bas
@@ -8,6 +8,6 @@
 
 #include "vbcompat.bi"
 
-Dim a As Double = DateSerial(2005, 11, 28)
+Dim a As Long = DateSerial(2005, 11, 28)
 
 Print Format(a, "yyyy/mm/dd hh:mm:ss")
