+++ examples/manual/dates/day.bas
@@ -8,6 +8,6 @@
 
 #include "vbcompat.bi"
 
-Dim ds As Double = DateSerial(2005, 11, 28)
+Dim ds As Long = DateSerial(2005, 11, 28)
 
 Print Format(ds, "yyyy/mm/dd "); Day(ds)
