+++ examples/manual/gfx/screencontrol.bas
@@ -13,7 +13,7 @@
 Using FB
 
 Dim e As Event
-Dim As Integer x0, y0, x, y
+Dim As Long x0, y0, x, y
 Dim As Integer shakes = 0
 Dim As Any Ptr img
