+++ examples/manual/strings/valint.bas
@@ -6,7 +6,7 @@
 '' See Also: https://www.freebasic.net/wiki/wikka.php?wakka=KeyPgValint
 '' --------
 
-Dim a As String, b As Integer
+Dim a As String, b As Long
 a = "20xa211"
 b = ValInt(a)
 Print a, b
