+++ examples/manual/system/kill.bas
@@ -7,6 +7,6 @@
 '' --------
 
 Dim filename As String = "file.ext"
-Dim result As Integer = Kill( filename )
+Dim result As Long = Kill( filename )
 
 If result <> 0 Then Print "error trying to kill " ; filename ; " !"
