+++ examples/manual/system/mkdir.bas
@@ -7,6 +7,6 @@
 '' --------
 
 Dim pathname As String = "foo\bar\baz"
-Dim result As Integer = MkDir( pathname )
+Dim result As Long = MkDir( pathname )
 
 If 0 <> result Then Print "error: unable to create folder " & pathname & " in the current path."
